Split useEntryState into directory
We intend to add more functionality.

REDMINE-17301
