+++ entry_types/scrolled/package/src/entryState/index.js
@@ -1,0 +1,2 @@
+export {useEntryState} from './useEntryState';
+export {watchCollections} from './watchCollections';
