+++ entry_types/scrolled/package/src/entryState/useEntryState.js
@@ -1,6 +1,6 @@
 import {useMemo} from 'react';
 
-import {useCollections, getItems, watchCollection} from './collections';
+import {useCollections, getItems} from '../collections';
 
 export function useEntryState(seed = {}) {
   const [collections, dispatch] = useCollections(seed.collections);
@@ -33,25 +33,3 @@ export function useEntryState(seed = {}) {
     dispatch
   ]
 };
-
-export function watchCollections({chapters, sections, contentElements}, {dispatch}) {
-  watchCollection(chapters, {
-    name: 'chapters',
-    attributes: ['id', 'permaId'],
-    includeConfiguration: true,
-    dispatch
-  });
-  watchCollection(sections, {
-    name: 'sections',
-    attributes: ['id', 'permaId', 'chapterId'],
-    includeConfiguration: true,
-    dispatch
-  });
-  watchCollection(contentElements, {
-    name: 'contentElements',
-    attributes: ['id', 'permaId', 'typeName', 'sectionId'],
-    keyAttribute: 'permaId',
-    includeConfiguration: true,
-    dispatch
-  });
-}
